upgraded to use Big Integer

+++ backend/src/main/java/com/health_insurance/backend/model/ClaimHistory.java
@@ -8,6 +8,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 @Entity
@@ -33,7 +34,7 @@ public class ClaimHistory {
     private BigDecimal amountPaid;
 
     @Column(name = "claimPersonaID", nullable = false)
-    private Long claimPersonaID;
+    private BigInteger claimPersonaID;
 
     @Temporal(TemporalType.TIMESTAMP)
     @Column(name = "timeStamp", nullable = false, updatable = false, columnDefinition = "default CURRENT_TIMESTAMP")
